Add install instructions, move workflow and config to root
